Inline the code from blob-stream and make it web compatible
Adaptations taken from https://github.com/devongovett/blob-stream/pull/7
which is unlikely to ever get merged, judging from the fact that the
package has not been updated since 2014.

blob-stream is only used in the web version of react-pdf, so it is safe
to directly use stream-browserify, because it will never be used on
node.